july 14 changes
box colliders, beach png line up, replaced some clouds, resized image of jump and duck buttons but kept collider sizes of them, replaced the greek pngs to fix platform height. moved seagull and crab pair a little to the right in one of the scenes it was difficult to avoid them.